requirements.txt
line-bot-sdk==3.11.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
apscheduler==3.10.4
Pillow==10.4.0
aiosqlite==0.20.0
sqlalchemy==2.0.35
